meta.setify: Fixed a typo in ddoc example.

+++ std/internal/meta/meta.d
@@ -5064,7 +5064,7 @@ Example:
 alias meta.setify!(int, bool, double, int) A;
 alias meta.setify!(bool, bool, double, int) B;
 
-static assert(is(A == TypeSeq!(bool, int, double, int)));
+static assert(is(A == TypeSeq!(bool, double, int, int)));
 static assert(is(B == TypeSeq!(bool, bool, double, int)));
 
 // Use meta.uniq to ignore the duplicates.
